requirements.txt
numpy~=1.19.2
matplotlib~=3.3.2
torch~=1.8.0
MiniSom~=2.2.9
scikit-learn~=1.1.1
scikit-learn-intelex~=2025.5.0
RecForest~=0.1.0
tqdm~=4.51.0
pandas~=1.2.1
joblib~=1.0.0
scipy~=1.6.0
setuptools~=56.1.0
umap-learn
